refactor: 패키지 구조 변경
- 관심사 기준으로 패키징합니다.
