+++ src/main/java/model/ladder/Bridge.java
@@ -1,4 +1,4 @@
-package model;
+package model.ladder;
 
 public enum Bridge {
     EMPTY, EXIST;
