+++ src/test/java/model/PlayerTest.java
@@ -3,6 +3,7 @@ package model;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import model.players.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
